requirements.txt
langchain
langchain-community
langchain-openai
rank_bm25
streamlit
python-dotenv
redis
